// src/app/core/interfaces/iarticle.ts
export interface IArticle {
  title: string;
  author: string;
  published_Date: string;
  published_Date_Precision: string;
  link: string;
  clean_Url: string;
  excerpt: string;
  summary: string;
  rights: string;
  rank: number;
  topic: string;
  country: string;
  language: string;
  authors: string[];
  media: string;
  is_Opinion: boolean;
  twitter_Account: string;
  _Id: string;
}